refactor: add inference code
