+++ .gitignore
@@ -163,3 +163,4 @@ cython_debug/
 *test.py
 
 lightning_logs/
+*.wav
